Bugfix: Update task-list title correctly
When the title of a task-list is updated, the corresponding key in the
state's `task-lists` was not updated. As a result, after the user
finishes editing the title of the task list, the task list items would
disappear because the updated title is not present in the state's
`task-lists` entry.

This commit fixes the issue by updating the task-list title correctly,
by using an intermediate title-buffer in the state